Continuing cleanup work on the Backdrop port. issue #1: Suggested fixes by herbdool
I chose to leave the code in that utilizes the php filter module for now,
until we feel sure it is never going to re-surface as a contrib module.  If
not, it should eventually be removed.  There has been a lot of discussion
on the issue.  https://www.drupal.org/project/drupal/issues/1203886

All this commit does is add a detailed comment about it in
css_injector.admin.inc.